Fixed the weird collisions with the bridge
Issue: Players would occassionally collide with the bridge
in the MainGame scene.

Why did it happen: It was happening because the offset for the
top of the collider in ShouldPassThrough was not being computed
correctly due to an error in localScale. localScale only refers to
the scale in relation to the immediate parent, but not any scale
that might exist above that.

The fix: Using Transform.TransformVector properly transformed the
offset vector into worldspace
was used