exp fix for double bullet and bullet latency
since bullets are being simulated locally we give each local copy
of the bullet a chance to damage the player locally (i.e. damage/death)
is no longer entirely one sided ("as soon as one client sees it,
everyone sees it immediately"). If after a bullet strikes another
player on machine A, that same bullet strikes that player on machine
B then we inflict damage for both machine A and B at the team that
they would see the bullet hit the player. To ensure consistency though
if client A sees a bullet hit a player and client B does not see that
within some small window of time, we correct client B and force the
damage and bullet despawn through