fixes #441
my old method for filtering out non-submit end-edits for the input
field was convoluted and ultimately didn't work. This was because when
the dynamic ui input system was initially deselecting the event systems
first selected object in mouse mode it was triggering the OnEndEdit
callback which was causing all kinds of problems. Now we just check if
submit was pressed anytime the OnEndEdit callback is called and if it
wasn't we return early